requirements.txt
fastapi
uvicorn
requests
python-telegram-bot==20.0
